Reorganized and consolidated types and models related to the parkView. There was some type duplication between the workspaceApi State model and the ParkView Models. This made updating types unnecessarily bothersome and had already caused differencens between seemingly similar types. Models related to ParkView is now moved to the Parkview component, and removed from the workspaceApi. Trying to "untie" the relation between workspace and ParkView. No reason for parkView to be "beholden" to the workspaceApi in anyway.
