+++ src/Core/WorkSpace/src/WorkSpaceApi/State.ts
@@ -4,8 +4,11 @@ import { FilterOptions } from '@equinor/filter';
 import { CustomCell, CustomColumn, CustomHeader } from '@equinor/Table';
 import React from 'react';
 import { TableOptions as ReactTableOptions } from 'react-table';
-import { FieldSettings } from '../../../../apps/swcr/models/FieldSettings';
-import { Data, DataSet } from '../../../../components/ParkView/Models/data';
+import {
+    CustomView,
+    GardenOptions,
+    StatusView,
+} from '../../../../components/ParkView/Models/gardenOptions';
 import { Filter } from '../../../../modules/powerBI/src/models/filter';
 import { StatusItem } from '../../../../packages/StatusBar';
 import { DataSource, DataViewerProps, ViewOptions } from './WorkSpaceTypes';
@@ -40,57 +43,14 @@ export interface Status {
     status?: string;
 }
 
-interface StatusView<T> {
-    statusItemFunc: (data: T) => Status;
-    statusGroupFunc?: (group: DataSet<T>) => Status;
-    shouldAggregate: boolean;
-}
-
 interface Options<T> {
     groupDescriptionFunc?: (data: T, groupingKey: string) => string;
-}
-
-export interface CustomItemView<T> {
-    data: T;
-    itemKey: string;
-    onClick: () => void;
-    columnExpanded: boolean;
-}
-
-export interface CustomGroupView<T> {
-    data: DataSet<T>;
-    onClick: () => void;
-    columnExpanded: boolean;
-}
-
-export interface CustomHeaderView<T> {
-    garden: Data<T>;
-    columnKey: string;
-}
-
-interface CustomView<T> {
-    customItemView?: React.FC<CustomItemView<T>>;
-    customGroupView?: React.FC<CustomGroupView<T>>;
-    customHeaderView?: React.FC<CustomHeaderView<T>>;
 }
 
 //update TreeOptions;;
 export interface TreeOptions<T> {
     groupByKeys?: (keyof T)[];
     itemKey: keyof T;
-    excludeKeys?: (keyof T)[];
-    customViews?: CustomView<T>;
-    options?: Options<T>;
-    status?: StatusView<T>;
-}
-
-export type GetKeyFunction<T> = (item: T, itemKey?: keyof T) => string[];
-
-export interface GardenOptions<T> {
-    gardenKey: keyof T;
-    itemKey: keyof T;
-    groupByKeys?: (keyof T)[];
-    fieldSettings?: FieldSettings<T, string>;
     excludeKeys?: (keyof T)[];
     customViews?: CustomView<T>;
     options?: Options<T>;
