+++ src/components/ParkView/Context/ParkViewContext.ts
@@ -1,8 +1,7 @@
 import { createContext } from 'react';
-import { FieldSettings } from '../../../apps/swcr/models/FieldSettings';
-
-import { GardenOptions, Status, TreeOptions } from '../../../Core/WorkSpace/src/WorkSpaceApi/State';
-import { Data, DataSet } from '../Models/data';
+import { TreeOptions } from '../../../Core/WorkSpace/src/WorkSpaceApi/State';
+import { DataSet } from '../Models/data';
+import { GardenOptions } from '../Models/gardenOptions';
 
 export interface CustomItemViewProps<T> {
     data: T;
@@ -15,44 +14,11 @@ export interface CustomGroupViewProps<T> {
     onClick: () => void;
 }
 
-export interface Options<T> {
-    groupDescriptionFunc?: (data: T, groupingKey: string) => string;
+export interface ParkViewState<T> extends GardenOptions<T> {
+    data: T[] | undefined;
 }
 
-export interface StatusView<T> {
-    statusItemFunc: (data: T) => Status;
-    statusGroupFunc?: (group: DataSet<T>) => Status;
-    shouldAggregate: boolean;
-}
-
-export interface CustomView<T> {
-    customItemView?: React.FC<{
-        data: T;
-        itemKey: string;
-        onClick: () => void;
-        columnExpanded: boolean;
-    }>;
-    customGroupView?: React.FC<{
-        data: DataSet<any>;
-        onClick: () => void;
-        columnExpanded: boolean;
-    }>;
-    customHeaderView?: React.FC<{ garden: Data<T>; columnKey: string }>;
-}
-
-export interface ParkViewState {
-    gardenKey?: string;
-    itemKey: string;
-    groupByKeys: string[];
-    excludeKeys?: string[];
-    fieldSettings?: FieldSettings<unknown, string>;
-    customView: CustomView<unknown>;
-    options?: Options<unknown>;
-    status?: StatusView<unknown>;
-    data: unknown[] | undefined;
-}
-
-export interface ParkViewContextState extends ParkViewState {
+export interface ParkViewContextState extends ParkViewState<unknown> {
     setGroupKeys: (groupKeys: string[]) => void;
     setExcludeKeys: (excludeKeys: string[]) => void;
     setGardenKey: (groupeKey?: string) => void;
