+++ src/components/ParkView/Context/ParkViewProvider.tsx
@@ -1,15 +1,9 @@
 import { useContext, useEffect, useReducer } from 'react';
-import { FieldSettings } from '../../../apps/swcr/models/FieldSettings';
-import { GardenOptions } from '../../../Core/WorkSpace/src/WorkSpaceApi/State';
+import { FieldSettings } from '../Models/FieldSettings';
+import { CustomView, GardenOptions, Options, StatusView } from '../Models/gardenOptions';
+
 import { actions } from './ParkViewActions';
-import {
-    CustomView,
-    Options,
-    ParkViewContext,
-    ParkViewProviderProps,
-    ParkViewState,
-    StatusView,
-} from './ParkViewContext';
+import { ParkViewContext, ParkViewProviderProps, ParkViewState } from './ParkViewContext';
 import { GardenReducer } from './ParkViewReducer';
 
 export function ParkViewProvider<T>({
@@ -17,16 +11,10 @@ export function ParkViewProvider<T>({
     data,
     parkViewOptions,
 }: ParkViewProviderProps<T>): JSX.Element {
-    const initialState: ParkViewState = {
-        itemKey: parkViewOptions.itemKey.toString(),
-        excludeKeys: parkViewOptions.excludeKeys as string[],
-        groupByKeys: (parkViewOptions.groupByKeys as string[]) || [],
-        customView: parkViewOptions.customViews as CustomView<unknown>,
-        options: parkViewOptions.options as Options<unknown>,
-        status: parkViewOptions.status as StatusView<unknown>,
-        fieldSettings: (parkViewOptions as GardenOptions<unknown>)?.fieldSettings || {},
+    const initialState: ParkViewState<T> = {
+        ...parkViewOptions,
         data: data,
-        gardenKey: (parkViewOptions as GardenOptions<T>)?.gardenKey?.toString(),
+        gardenKey: (parkViewOptions as GardenOptions<T>)?.gardenKey,
     };
     const [state, dispatch] = useReducer(GardenReducer, initialState);
 
@@ -72,7 +60,7 @@ export function useParkViewContext<T>() {
         itemKey: parkViewContext.itemKey as keyof T,
         excludeKeys: parkViewContext.excludeKeys as (keyof T)[],
         groupByKeys: parkViewContext.groupByKeys as (keyof T)[],
-        customView: parkViewContext.customView as CustomView<T>,
+        customView: parkViewContext.customViews as CustomView<T>,
         status: parkViewContext.status as StatusView<T>,
         options: parkViewContext.options as Options<T>,
         data: parkViewContext.data as T[],
