+++ src/components/ParkView/Models/fieldSettings.ts
@@ -1,6 +1,5 @@
-import { GetKeyFunction } from '../../../Core/WorkSpace/src/WorkSpaceApi/State';
-
 export type GetSortFunction = (a: string, b: string) => number;
+export type GetKeyFunction<T> = (item: T, itemKey?: keyof T) => string[];
 
 export type FieldSetting<ItemType> = {
     key?: keyof ItemType;
